Незаконченная работа

+++ src/main/java/ru/skypro/homework/dto/UserDto.java
@@ -22,7 +22,7 @@ public class UserDto {
   private String firstName = null;
 
   @JsonProperty("id")
-  private Integer id = null;
+  private Long id = null;
 
   @JsonProperty("lastName")
   private String lastName = null;
@@ -68,7 +68,7 @@ public class UserDto {
     this.firstName = firstName;
   }
 
-  public UserDto id( Integer id) {
+  public UserDto id( Long id) {
     this.id = id;
     return this;
   }
@@ -79,11 +79,11 @@ public class UserDto {
    **/
   @Schema(description = "")
   
-    public Integer getId() {
+    public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
